requirements.txt
Flask==1.1.4
Jinja2==2.11.3
markupsafe==1.1.1
python-telegram-bot==13.7
requests==2.25.1
python-dotenv==0.19.0
SQLAlchemy==1.4.22
apscheduler==3.6.3
pytz==2021.1
Werkzeug==1.0.1

